NP-285 clearing up Json literals and static code
