--- app/uk/gov/hmrc/uknwauthcheckerapistub/models/requests/Requests200.scala
@@ -1,24 +1,0 @@
-/*
- * Copyright 2024 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.uknwauthcheckerapistub.models.requests
-
-object Requests200 {
-  val req200_single: Seq[String] = Seq("GB000000000200")
-
-  val req200_multiple: Seq[String] = Seq("GB000000000200", "XI000000000200")
-
-}
