+++ app/uk/gov/hmrc/uknwauthcheckerapistub/utils/makers/OkMaker.scala
@@ -1,0 +1,9 @@
+package uk.gov.hmrc.uknwauthcheckerapistub.utils.makers
+
+import uk.gov.hmrc.uknwauthcheckerapistub.models.EoriResults
+
+class OkMaker {
+  
+  def makeOk(eoris: Seq[String]): Seq[EoriResults] = ???
+
+}
