+++ it/test/uk/gov/hmrc/uknwauthcheckerapistub/EisStubControllerISpec.scala
@@ -17,7 +17,6 @@
 package uk.gov.hmrc.uknwauthcheckerapistub
 
 import play.api.http.Status
-import uk.gov.hmrc.uknwauthcheckerapistub.models.requests.PerformanceRequests._
 import uk.gov.hmrc.uknwauthcheckerapistub.models.requests.Requests200._
 
 class EisStubControllerISpec extends BaseISpec {
