+++ test/uk/gov/hmrc/uknwauthcheckerapistub/controllers/TestDataUtils.scala
@@ -17,7 +17,6 @@
 package uk.gov.hmrc.uknwauthcheckerapistub.controllers
 
 import play.api.libs.json.{JsValue, Json}
-import uk.gov.hmrc.uknwauthcheckerapistub.utils.JsonGetter
 
 trait TestDataUtils extends JsonGetter {
